Don't quit driver if environment contains TP_KEEP_DRIVER_SESSION variable
OpenSDK has logic that closes the driver when process exits.
However we cannot quit the driver if executing from the platform, because that will break other tests.
To prevent calls to driver.Quit() from killing the session, agent specifies an environment variable.
SDK should respect this flag and not quit the session if it is set

Signed-off-by: shai glatshtein <shai.glatshtein@testproject.io>